Refactor spellcaster routes
Implement fast path: imagine
